+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Enable type checking on methods `evaluate`, `evaluateWithReturn` and `waitForEvaluation`.
 
 ## [0.8.0] - 2018-08-23
 ### Changed